refactor: isolate install code in agent/install, move marker to connector
- Move the per-project marker (.codesearch/project.json read/write/find) out of
  the agent module into the connector layer (src/connector/marker.rs), where
  filesystem persistence belongs; it is consumed by both the index controller
  and the pre-tool-call command.
- Group all install/uninstall/hook-writing code under a dedicated
  src/agent/install submodule, leaving src/agent with just the install
  subtree and the runtime pre_tool_call command.

Repoints references (main.rs, connector/mod.rs, index_controller, pre_tool_call).
No behaviour change.
