+++ src/agent/install/opencode.rs
