+++ src/agent/install/zed.rs
