+++ src/connector/mod.rs
@@ -1,4 +1,5 @@
 pub mod adapter;
 pub mod api;
+pub mod marker;
 
 pub use adapter::*;
